refactor(dashboard): route ui deep imports through barrel
Six layout files (app-shell, command-palette, header, mobile-menu,
route-error-boundary, theme-toggle) now import from '@/components/ui'
instead of per-file paths, matching the pattern used in route files.

routes/jobs/index.tsx had three separate '@/features/jobs/*' imports
despite already using the feature barrel; collapses JobFilters and
JobListQuery type imports into the existing barrel import.

providers/index.tsx keeps its '@/components/ui/toaster' deep import
to keep the providers shell lean. routes/logs.tsx and routes/metrics.tsx
keep theirs per file-level comments (bundle-splitting: the non-lazy
route must not pull Recharts or virtualized-log components).
