+++ dashboard/src/components/layout/header.tsx
@@ -1,6 +1,5 @@
 import { Search } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { Kbd } from "@/components/ui/kbd";
+import { Button, Kbd } from "@/components/ui";
 import { useCommandPalette } from "@/providers";
 import { LastRefreshed } from "./last-refreshed";
 import { MobileMenu } from "./mobile-menu";
