+++ dashboard/src/components/layout/mobile-menu.tsx
@@ -14,8 +14,14 @@ import {
   Skull,
 } from "lucide-react";
 import { useEffect, useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Sheet, SheetContent, SheetHeader, SheetTitle, SheetTrigger } from "@/components/ui/sheet";
+import {
+  Button,
+  Sheet,
+  SheetContent,
+  SheetHeader,
+  SheetTitle,
+  SheetTrigger,
+} from "@/components/ui";
 import { cn } from "@/lib/cn";
 import { site } from "@/lib/site";
