+++ dashboard/src/components/layout/theme-toggle.tsx
@@ -1,5 +1,5 @@
 import { Laptop, Moon, Sun } from "lucide-react";
-import { Button } from "@/components/ui/button";
+import { Button } from "@/components/ui";
 import { cn } from "@/lib/cn";
 import { type Theme, useTheme } from "@/providers";
